+++ particle_tracking_manager/__init__.py
@@ -12,6 +12,7 @@ logging.getLogger("numcodecs").setLevel(logging.WARNING)
 from .config_the_manager import TheManagerConfig
 from .models import opendrift as opendrift_models
 from .models.opendrift.config_opendrift import (
+    HarmfulAlgalBloomModelConfig,
     LarvalFishModelConfig,
     LeewayModelConfig,
     OceanDriftModelConfig,
@@ -29,5 +30,6 @@ __all__ = [
     "OpenDriftConfig",
     "OpenOilModelConfig",
     "OpenDriftModel",
+    "HarmfulAlgalBloomModelConfig",
     "opendrift_models",
 ]