+++ particle_tracking_manager/models/opendrift/enums/__init__.py
@@ -1,7 +1,7 @@
 """Enums for OpenDrift."""
 
 from .object_types import ObjectTypeEnum
-from .oil_types import OIL_ID_TO_NAME, OilTypeEnum
+from .oil_types import NAME_TO_OIL_ID, OIL_ID_TO_NAME, OilTypeEnum
 from .others import (
     CoastlineActionEnum,
     DiffusivityModelEnum,
@@ -11,17 +11,20 @@ from .others import (
     RadiusTypeEnum,
     SeafloorActionEnum,
 )
+from .species_types import HABSpeciesTypeEnum
 
 
 __all__ = [
     "OilTypeEnum",
     "OIL_ID_TO_NAME",
+    "NAME_TO_OIL_ID",
     "ObjectTypeEnum",
     "DriftModelEnum",
     "RadiusTypeEnum",
     "DiffusivityModelEnum",
     "CoastlineActionEnum",
     "SeafloorActionEnum",
+    "HABSpeciesTypeEnum",
     "PlotTypeEnum",
     "DropletSizeDistributionEnum",
 ]